#44 Refactor buy to recieve button on dashboard
